pyproject.toml
[project]
name = "Kabuda-Backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "PookieBuns", email = "allenyuchen@hotmail.com"},
]
dependencies = ["fastapi>=0.115.8", "uvicorn>=0.34.0", "pydantic-settings>=2.7.1", "loguru>=0.7.3", "PyJWT>=2.10.1", "arrow>=1.3.0", "pydantic>=2.10.6"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[dependency-groups]
lint = [
    "mypy>=1.15.0",
]
format = [
    "black>=25.1.0",
    "isort>=6.0.0",
]

[tool.mypy]
plugins = ['pydantic.mypy']

[tool.pdm]
distribution = false

